fix conversion

+++ tests/online_judge_tests/math/matrix_determinant.test.cpp
@@ -7,7 +7,7 @@ int main() {
 	cin.tie(0)->sync_with_stdio(0);
 	int n;
 	cin >> n;
-	vector<vector<int>> matrix(n, vector<int>(n));
+	vector<vector<long long>> matrix(n, vector<long long>(n));
 	for (int i = 0; i < n; i++) {
 		for (int j = 0; j < n; j++)
 			cin >> matrix[i][j];
